Fix: Entity 리팩토링 컨벤션을 고려하여 엔티티 컬럼을 수정함

+++ modac/src/main/java/com/a608/modac/model/user/User.java
@@ -25,7 +25,7 @@ import lombok.ToString;
 @Getter
 @Entity
 @ToString
-@DynamicInsert	//not null인 컬럼에 데이터 저장 안해줄때 자동으로 default값으로 매핑해줌
+@DynamicInsert    //not null인 컬럼에 데이터 저장 안해줄때 자동으로 default값으로 매핑해줌
 @NoArgsConstructor
 @Table(name = "users")
 public class User implements UserDetails {
@@ -37,10 +37,10 @@ public class User implements UserDetails {
 	@Column(name = "id")
 	private String id;
 
-	@Column(name="nickname")
+	@Column(name = "nickname")
 	private String nickname;
 
-	@Column(name="email")
+	@Column(name = "email")
 	private String email;
 
 	@Column(name = "password")
@@ -53,7 +53,7 @@ public class User implements UserDetails {
 	private String singleTheme;    //싱글룸 테마(우주....)
 
 	@Column(name = "total_second")
-	private Long totalSecond; //인생 총 공부시간
+	private Integer totalSecond; //인생 총 공부시간
 
 	@Column(name = "membership_level")
 	private String membershipLevel;
@@ -66,7 +66,7 @@ public class User implements UserDetails {
 
 	@Builder
 	public User(Long seq, String id, String nickname, String email, String password, Byte catSkin, String singleTheme,
-		Long totalSecond) {
+		Integer totalSecond) {
 		this.seq = seq;
 		this.id = id;
 		this.nickname = nickname;
@@ -77,7 +77,7 @@ public class User implements UserDetails {
 		this.totalSecond = totalSecond;
 	}
 
-	public void updateUser(String nickname, String email){
+	public void updateUser(String nickname, String email) {
 		this.nickname = nickname;
 		this.email = email;
 	}
